Fix: change int to bigint in district migration

+++ db/migrate/20221018035723_create_districts.rb
@@ -2,26 +2,26 @@ class CreateDistricts < ActiveRecord::Migration[5.2]
   def change
     create_table :districts do |t|
       t.string :name
-      t.integer :lea_id
+      t.bigint :lea_id
       t.string :urban_centric_locale
-      t.integer :number_of_schools_in_district
-      t.integer :enrollment
-      t.integer :pre_k_teachers
-      t.integer :kindergarten_teachers
-      t.integer :elementary_teachers
-      t.integer :secondary_teachers
-      t.integer :total_teachers
+      t.bigint :number_of_schools_in_district
+      t.bigint :enrollment
+      t.bigint :pre_k_teachers
+      t.bigint :kindergarten_teachers
+      t.bigint :elementary_teachers
+      t.bigint :secondary_teachers
+      t.bigint :total_teachers
       t.float :student_teacher_ratio
-      t.integer :instructional_aides
-      t.integer :guidance_counselors
+      t.bigint :instructional_aides
+      t.bigint :guidance_counselors
       t.float :student_guidance_counselor_ratio
-      t.integer :librarian_total
+      t.bigint :librarian_total
       t.float :student_librarian_ratio
-      t.integer :total_staff
-      t.integer :district_expense_total
-      t.integer :expenses_for_instruction
-      t.integer :salaries_total
-      t.integer :salaries_instruction
+      t.bigint :total_staff
+      t.bigint :district_expense_total
+      t.bigint :expenses_for_instruction
+      t.bigint :salaries_total
+      t.bigint :salaries_instruction
       t.float :instruction_salary_percent_of_total
       t.float :per_student_expenditure
       t.float :per_teacher_salary_expenses
